Add a hack to share structs
