+++ src-tauri/src/main.rs
@@ -10,7 +10,6 @@ use anyhow::{Context as _, Result, anyhow, bail};
 
 use komorebi_client::{Window, Workspace};
 use komorebi_client::{send_query, SocketMessage, State};
-use serde::{Deserialize, Serialize};
 use tauri::Manager as _;
 use tauri::SystemTray;
 use tauri::SystemTrayEvent;
@@ -121,32 +120,8 @@ fn fetch_komorebi_state() -> Result<State> {
     // or, you may not running komorebi
 }
 
-#[derive(Debug, PartialEq, Eq, Deserialize, Serialize)]
-struct AsayakeMonitorState {
-    monitor_index: usize,
-    focusing_workspace: usize,
-    workspaces: Vec<WorkspaceForSend>,
-}
-
-#[derive(Debug, PartialEq, Eq, Deserialize, Serialize)]
-struct WorkspaceForSend {}
-
-#[derive(Debug, PartialEq, Eq, Deserialize, Serialize)]
-enum WorkspaceItem {
-    Window(WindowForSend),
-    WindowStack(Vec<WindowForSend>)
-}
-
-#[derive(Debug, PartialEq, Eq, Deserialize, Serialize)]
-struct WindowForSend {
-    icon: Icon,
-    accent_color: u32
-}
-
-#[derive(Debug, PartialEq, Eq, Deserialize, Serialize)]
-struct Icon {
-    base64_icon: String,
-}
+// HACK: クレートを分割するとimplできないという制約を乗り越えるための方法。醜い
+include!("./structs.rs");
 
 impl From<&Workspace> for WorkspaceForSend {
     fn from(value: &Workspace) -> Self {
